Добавление в библиотеку Callback-модуля.

+++ python_rucaptcha/MediaCaptcha.py
@@ -84,7 +84,15 @@ class MediaCaptcha:
                             скриптом.
         :param audio_download_link: Передаётся ссылка для скачивания аудио файла. Не ссылка на капчу или ещё что-либо.
                                     А именно ссылка по которой можно скачать аудио файл. Для последующей отправке RuCaptcha.
-        :return: Возвращает решение капчи.
+		:return: Ответ на капчу в виде JSON строки с полями:
+                    captchaSolve - решение капчи,
+                    taskId - находится Id задачи на решение капчи, можно использовать при жалобах и прочем,
+                    error - False - если всё хорошо, True - если есть ошибка,
+                    errorBody - полная информация об ошибке:
+                        {
+                            text - Развернётое пояснение ошибки
+                            id - уникальный номер ошибка в ЭТОЙ бибилотеке
+                        }
         """
         # результат возвращаемый методом *captcha_handler*
         self.result = JSON_RESPONSE.copy()
@@ -131,12 +139,17 @@ class MediaCaptcha:
             self.result.update({"taskId": captcha_id})
             # обновляем пайлоад, вносим в него ключ отправленной на решение капчи
             self.get_payload.update({'id': captcha_id})
-
-        # удаляем файл капчи
-        os.remove(os.path.join(self.audio_path, f'aud-{audio_hash}.mp3'))
-        # Ожидаем решения капчи
-        time.sleep(self.sleep_time)
-        return get_sync_result(get_payload=self.get_payload,
-                               sleep_time = self.sleep_time,
-                               url_response = self.url_response,
-                               result = self.result)
+            
+            # если передан параметр `pingback` - не ждём решения капчи а возвращаем незаполненный ответ
+            if self.post_payload.get('pingback'):
+                return self.get_payload
+            
+            else:
+                # удаляем файл капчи
+                os.remove(os.path.join(self.audio_path, f'aud-{audio_hash}.mp3'))
+                # Ожидаем решения капчи
+                time.sleep(self.sleep_time)
+                return get_sync_result(get_payload=self.get_payload,
+                                    sleep_time = self.sleep_time,
+                                    url_response = self.url_response,
+                                    result = self.result)
